LayoutTests/fast/events/message-channel-gc-expected.txt

CONSOLE MESSAGE: line 25: ReferenceError: Can't find variable: MessageChannel
FAIL: Timed out waiting for notifyDone to be called
Test some cases of MessagePort garbage collection.
